Refactored Group & Lock Controllers

+++ app/Http/Controllers/LockController.php
@@ -6,13 +6,24 @@ use App\Device;
 use App\Helpers\ResponseWrapper;
 use App\Http\Resources\DeviceResource;
 use App\Lock;
+use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Helpers\LogHelper;
 use App\Http\Resources\LockResource;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Date;
 
+/**
+ * Class LockController
+ * @package App\Http\Controllers
+ */
 class LockController extends Controller
 {
+    /**
+     * LockController constructor. Applies API-Key authentication middleware
+     */
     public function __construct()
     {
         $this->middleware('key.auth');
@@ -20,7 +31,7 @@ class LockController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection|\Illuminate\Http\Response
+     * @return AnonymousResourceCollection|Response
      */
     public function index()
     {
@@ -31,7 +42,7 @@ class LockController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
@@ -41,11 +52,16 @@ class LockController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return LockResource|\Illuminate\Http\Response
+     * @param Request $request
+     * @return LockResource|JsonResponse|Response
      */
     public function store(Request $request)
     {
+        if(!$request->filled('name'))
+            return ResponseWrapper::wrap('Name field missing', $request->all(), ResponseWrapper::BAD_REQUEST);
+        if(!$request->filled('category'))
+            return ResponseWrapper::wrap('Category field missing', $request->all(), ResponseWrapper::BAD_REQUEST);
+
         $lock = new Lock();
         $lock->name = $request->input('name');
         $lock->status = $request->input('status');
@@ -53,15 +69,15 @@ class LockController extends Controller
         if($lock->save()){
             LogHelper::Log($request->input('user_id'), $lock, LogHelper::Lock, "Store");
             return new LockResource($lock);
-
         }
+        return ResponseWrapper::wrap('Lock not saved', $request->all(), ResponseWrapper::SERVER_ERROR);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Lock  $lock
-     * @return LockResource|\Illuminate\Http\Response
+     * @param Lock $lock
+     * @return LockResource|Response
      */
     public function show(Lock $lock)
     {
@@ -71,8 +87,8 @@ class LockController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Lock  $lock
-     * @return \Illuminate\Http\Response
+     * @param Lock $lock
+     * @return Response
      */
     public function edit(Lock $lock)
     {
@@ -82,13 +98,14 @@ class LockController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Lock  $lock
-     * @return LockResource|\Illuminate\Http\Response
+     * @param Request $request
+     * @param Lock $lock
+     * @return LockResource|JsonResponse|Response
      */
     public function update(Request $request, Lock $lock)
     {
-
+        if(!$lock->exists)
+            return ResponseWrapper::wrap('Lock not found', request()->all(), ResponseWrapper::NOT_FOUND);
         $lock->name = $request->filled('name')? $request->input('name') : $lock->name;
         $lock->status = $request->filled('status') ? $request->input('status') : $lock->status;
 
@@ -96,21 +113,25 @@ class LockController extends Controller
             LogHelper::Log($request->input('user_id'), $lock, LogHelper::Lock, "Update");
             return new LockResource($lock);
         }
+        return ResponseWrapper::wrap('Device not updated', $request->all(), ResponseWrapper::SERVER_ERROR);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param \App\Lock $lock
-     * @return \Illuminate\Http\Response|string
-     * @throws \Exception
+     * @param Lock $lock
+     * @return Response|string
+     * @throws Exception
      */
     public function destroy(Lock $lock)
     {
+        if(!$lock->exists)
+            return ResponseWrapper::wrap('Device not found', request()->all(), ResponseWrapper::NOT_FOUND);
         if($lock->delete()){
             LogHelper::Log(request()->input('user_id'), $lock, LogHelper::Lock, "Destroy");
             return "Lock deleted.";
         }
+        return ResponseWrapper::wrap('Lock not deleted', request()->all(), ResponseWrapper::SERVER_ERROR);
     }
 
     public function keepAlive(Request $request, Device $device){
